+++ packages/react-components/src/components/ChatMessage/ChatMessageComponentWrapper.tsx
@@ -12,10 +12,12 @@ import { FileDownloadHandler } from '../FileDownloadCards';
 /* @conditional-compile-remove(mention) */
 import { MentionOptions } from '../MentionPopover';
 import { MessageStatusIndicatorProps } from '../MessageStatusIndicator';
-import { FluentChatMessageComponentWrapper } from './FluentChatMessageComponentWrapper';
+import { FluentChatMessageComponentWrapperProps } from './MessageComponents/FluentChatMessageComponent';
 import { DefaultSystemMessage } from './DefaultSystemMessage';
 /* @conditional-compile-remove(image-overlay) */
 import { InlineImageOptions } from './ChatMessageContent';
+import { FluentChatMyMessageComponent } from './MyMessageComponents/FluentChatMyMessageComponent';
+import { FluentChatMessageComponent } from './MessageComponents/FluentChatMessageComponent';
 
 /**
  * Props for {@link ChatMessageComponentWrapper}
@@ -89,9 +91,7 @@ export const ChatMessageComponentWrapper = (props: ChatMessageComponentWrapperPr
         : styles?.myChatMessageContainer;
     const blockedMessageStyle = styles?.blockedMessageContainer;
     const messageContainerStyle = message.mine ? myChatMessageStyle : blockedMessageStyle;
-    return (
-      <FluentChatMessageComponentWrapper {...props} message={message} messageContainerStyle={messageContainerStyle} />
-    );
+    return fluentChatComponent({ ...props, message: message, messageContainerStyle: messageContainerStyle });
   }
 
   switch (message.messageType) {
@@ -102,9 +102,7 @@ export const ChatMessageComponentWrapper = (props: ChatMessageComponentWrapperPr
           : styles?.myChatMessageContainer;
       const chatMessageStyle = styles?.chatMessageContainer;
       const messageContainerStyle = message.mine ? myChatMessageStyle : chatMessageStyle;
-      return (
-        <FluentChatMessageComponentWrapper {...props} message={message} messageContainerStyle={messageContainerStyle} />
-      );
+      return fluentChatComponent({ ...props, message: message, messageContainerStyle: messageContainerStyle });
     }
 
     case 'system': {
@@ -131,5 +129,13 @@ export const ChatMessageComponentWrapper = (props: ChatMessageComponentWrapperPr
         </div>
       );
     }
+  }
+};
+
+const fluentChatComponent = (props: FluentChatMessageComponentWrapperProps): JSX.Element => {
+  if (props.message.mine === true) {
+    return <FluentChatMyMessageComponent {...props} />;
+  } else {
+    return <FluentChatMessageComponent {...props} />;
   }
 };
